requirement.txt
aniso8601==9.0.1
Appium-Python-Client==2.6.0
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
concurrent-log-handler==0.9.20
cryptography==36.0.1
ddt==1.4.4
docopt==0.6.2
Flask==2.0.3
Flask-HTTPAuth==4.6.0
Flask-JWT==0.3.2
Flask-RESTful==0.3.9
func-timeout==4.3.5
h11==0.13.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==0.10.0
loguru==0.6.0
MarkupSafe==2.1.0
mysql-connector==2.2.9
numpy==1.22.3
outcome==1.1.0
packaging==21.3
paho-mqtt==1.5.1
pandas==1.4.2
pbr==5.8.1
pluggy==0.13.1
poium==1.1.5
portalocker==2.5.1
py==1.11.0
pycparser==2.21
PyJWT==1.7.1
pykwalify==1.8.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
pytest==6.2.4
pytest-base-url==1.4.2
pytest-dependency==0.5.1
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-rerunfailures==9.1.1
python-box==5.4.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
stevedore==3.5.0
tavern==1.22.0
toml==0.10.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.8
webdriver-manager==3.8.2
Werkzeug==2.0.3
wsproto==1.1.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
